Codes/requirements.txt
PyPDF2
tika
python-dotenv
SQLAlchemy
camelot-py[cv]
fitz
mysqlclient